pyproject.toml
[tool.poetry]
name = "djangocleanarch"
version = "0.1.0"
description = "Django clean arch - Poc"
authors = ["Higor Monteiro <higorvmonteiro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
gunicorn = "^23.0.0"
pillow = "^10.4.0"
sentry-sdk = "^2.14.0"
psycopg2 = "^2.9.9"
boto3 = "^1.35.24"
whitenoise = "^6.7.0"
sendgrid = "^6.11.0"
python-decouple = "^3.8"
openai = "^1.47.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ruff = "^0.6.6"
django-extensions = "^3.2.3"
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
